requirements.txt
asgiref==3.7.2
cloudinary==1.36.0
Django==3.2.23
django-cloudinary-storage==0.3.0
djangorestframework==3.14.0
Pillow==8.2.0
pytz==2023.3.post1
sqlparse==0.4.4
